safety checkin - about halfway through refactoring
